requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
Keras==2.1.6
Markdown==2.6.11
numpy==1.14.3
protobuf==3.5.2.post1
python-mnist==0.5
PyYAML==3.12
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.23.4
Werkzeug==0.14.1
